// src/types/gen/CrossTab.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CrossTabGroup } from "./CrossTabGroup";
import type { Dimension } from "./Dimension";
import type { ReportRow } from "./ReportRow";

export type CrossTab = { questionId: number, dimension: Dimension, 
/**
 * Answer columns: options, scale points or themes.
 */
columns: Array<ReportRow>, groups: Array<CrossTabGroup>, };